move engine into its own crate
