+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
   "crates/data-structures",
   "crates/datastore",
   "crates/durability",
+  "crates/engine",
   "crates/execution",
   "crates/expr",
   "crates/guard",
@@ -132,6 +133,7 @@ spacetimedb-core = { path = "crates/core", version = "=2.4.1" }
 spacetimedb-data-structures = { path = "crates/data-structures", version = "=2.4.1" }
 spacetimedb-datastore = { path = "crates/datastore", version = "=2.4.1" }
 spacetimedb-durability = { path = "crates/durability", version = "=2.4.1" }
+spacetimedb-engine = { path = "crates/engine", version = "=2.4.1" }
 spacetimedb-execution = { path = "crates/execution", version = "=2.4.1" }
 spacetimedb-expr = { path = "crates/expr", version = "=2.4.1" }
 spacetimedb-guard = { path = "crates/guard", version = "=2.4.1" }
@@ -143,6 +145,7 @@ spacetimedb-pg = { path = "crates/pg", version = "=2.4.1" }
 spacetimedb-physical-plan = { path = "crates/physical-plan", version = "=2.4.1" }
 spacetimedb-primitives = { path = "crates/primitives", version = "=2.4.1" }
 spacetimedb-query = { path = "crates/query", version = "=2.4.1" }
+spacetimedb-runtime = { path = "crates/runtime", version = "=2.4.1" }
 spacetimedb-sats = { path = "crates/sats", version = "=2.4.1" }
 spacetimedb-schema = { path = "crates/schema", version = "=2.4.1" }
 spacetimedb-standalone = { path = "crates/standalone", version = "=2.4.1" }
@@ -152,7 +155,6 @@ spacetimedb-fs-utils = { path = "crates/fs-utils", version = "=2.4.1" }
 spacetimedb-snapshot = { path = "crates/snapshot", version = "=2.4.1" }
 spacetimedb-subscription = { path = "crates/subscription", version = "=2.4.1" }
 spacetimedb-query-builder = { path = "crates/query-builder", version = "=2.4.1" }
-spacetimedb-runtime = { path = "crates/runtime", version = "=2.4.1" }
 
 # Prevent `ahash` from pulling in `getrandom` by disabling default features.
 # Modules use `getrandom02` and we need to prevent an incompatible version
