+++ crates/core/Cargo.toml
@@ -22,13 +22,14 @@ spacetimedb-client-api-messages.workspace = true
 spacetimedb-commitlog.workspace = true
 spacetimedb-datastore.workspace = true
 spacetimedb-durability.workspace = true
+spacetimedb-engine.workspace = true
 spacetimedb-memory-usage.workspace = true
 spacetimedb-metrics.workspace = true
 spacetimedb-primitives.workspace = true
 spacetimedb-paths.workspace = true
 spacetimedb-physical-plan.workspace = true
 spacetimedb-query.workspace = true
-spacetimedb-runtime = { workspace = true, features = ["tokio"] }
+spacetimedb-runtime.workspace = true
 spacetimedb-sats = { workspace = true, features = ["serde"] }
 spacetimedb-schema.workspace = true
 spacetimedb-table.workspace = true
@@ -144,7 +145,7 @@ allow_loopback_http_for_tests = []
 # Enable timing for wasm ABI calls
 spacetimedb-wasm-instance-env-times = []
 # Enable test helpers and utils
-test = ["spacetimedb-commitlog/test", "spacetimedb-datastore/test"]
+test = ["spacetimedb-commitlog/test", "spacetimedb-datastore/test", "spacetimedb-engine/test"]
 # Perfmaps for profiling modules
 perfmap = []
 # Enables core pinning.
