+++ crates/core/src/sql/mod.rs
@@ -1,3 +1,2 @@
 pub mod ast;
 pub mod execute;
-pub mod parser;
