+++ crates/core/src/subscription/mod.rs
@@ -1,19 +1,16 @@
+use crate::error::DBError;
 use crate::subscription::websocket_building::{BuildableWebsocketFormat, RowListBuilder as _, RowListBuilderSource};
-use crate::{error::DBError, worker_metrics::WORKER_METRICS};
 use anyhow::Result;
 use metrics::QueryMetrics;
 use module_subscription_manager::Plan;
-use prometheus::IntCounter;
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 use spacetimedb_client_api_messages::websocket::common::ByteListLen as _;
 use spacetimedb_client_api_messages::websocket::v1::{self as ws_v1};
-use spacetimedb_datastore::{
-    db_metrics::DB_METRICS, execution_context::WorkloadType, locking_tx_datastore::datastore::MetricsRecorder,
-};
+pub use spacetimedb_engine::metrics::ExecutionCounters;
 use spacetimedb_execution::pipelined::ViewProject;
 use spacetimedb_execution::{pipelined::PipelinedProject, Datastore, DeltaStore};
 use spacetimedb_lib::identity::AuthCtx;
-use spacetimedb_lib::{metrics::ExecutionMetrics, Identity};
+use spacetimedb_lib::metrics::ExecutionMetrics;
 use spacetimedb_primitives::TableId;
 use spacetimedb_sats::bsatn::ToBsatn;
 use spacetimedb_sats::Serialize;
@@ -31,72 +28,6 @@ pub mod row_list_builder_pool;
 pub mod subscription;
 pub mod tx;
 pub mod websocket_building;
-
-#[derive(Debug)]
-pub struct ExecutionCounters {
-    rdb_num_index_seeks: IntCounter,
-    rdb_num_rows_scanned: IntCounter,
-    rdb_num_bytes_scanned: IntCounter,
-    rdb_num_bytes_written: IntCounter,
-    bytes_sent_to_clients: IntCounter,
-    delta_queries_matched: IntCounter,
-    delta_queries_evaluated: IntCounter,
-    duplicate_rows_evaluated: IntCounter,
-    duplicate_rows_sent: IntCounter,
-}
-
-impl ExecutionCounters {
-    pub fn new(workload: &WorkloadType, db: &Identity) -> Self {
-        Self {
-            rdb_num_index_seeks: DB_METRICS.rdb_num_index_seeks.with_label_values(workload, db),
-            rdb_num_rows_scanned: DB_METRICS.rdb_num_rows_scanned.with_label_values(workload, db),
-            rdb_num_bytes_scanned: DB_METRICS.rdb_num_bytes_scanned.with_label_values(workload, db),
-            rdb_num_bytes_written: DB_METRICS.rdb_num_bytes_written.with_label_values(workload, db),
-            bytes_sent_to_clients: WORKER_METRICS.bytes_sent_to_clients.with_label_values(workload, db),
-            delta_queries_matched: DB_METRICS.delta_queries_matched.with_label_values(db),
-            delta_queries_evaluated: DB_METRICS.delta_queries_evaluated.with_label_values(db),
-            duplicate_rows_evaluated: DB_METRICS.duplicate_rows_evaluated.with_label_values(db),
-            duplicate_rows_sent: DB_METRICS.duplicate_rows_sent.with_label_values(db),
-        }
-    }
-
-    /// Update the global system metrics with transaction-level execution metrics.
-    pub(crate) fn record(&self, metrics: &ExecutionMetrics) {
-        if metrics.index_seeks > 0 {
-            self.rdb_num_index_seeks.inc_by(metrics.index_seeks as u64);
-        }
-        if metrics.rows_scanned > 0 {
-            self.rdb_num_rows_scanned.inc_by(metrics.rows_scanned as u64);
-        }
-        if metrics.bytes_scanned > 0 {
-            self.rdb_num_bytes_scanned.inc_by(metrics.bytes_scanned as u64);
-        }
-        if metrics.bytes_written > 0 {
-            self.rdb_num_bytes_written.inc_by(metrics.bytes_written as u64);
-        }
-        if metrics.bytes_sent_to_clients > 0 {
-            self.bytes_sent_to_clients.inc_by(metrics.bytes_sent_to_clients as u64);
-        }
-        if metrics.delta_queries_matched > 0 {
-            self.delta_queries_matched.inc_by(metrics.delta_queries_matched);
-        }
-        if metrics.delta_queries_evaluated > 0 {
-            self.delta_queries_evaluated.inc_by(metrics.delta_queries_evaluated);
-        }
-        if metrics.duplicate_rows_evaluated > 0 {
-            self.duplicate_rows_evaluated.inc_by(metrics.duplicate_rows_evaluated);
-        }
-        if metrics.duplicate_rows_sent > 0 {
-            self.duplicate_rows_sent.inc_by(metrics.duplicate_rows_sent);
-        }
-    }
-}
-
-impl MetricsRecorder for ExecutionCounters {
-    fn record(&self, metrics: &ExecutionMetrics) {
-        self.record(metrics);
-    }
-}
 
 /// Execute a subscription query over a view.
 ///
