+++ crates/core/src/worker_metrics/mod.rs
@@ -352,11 +352,6 @@ metrics_group!(
         #[labels(db: Identity, reducer: str)]
         pub reducer_plus_query_duration: HistogramVec,
 
-        #[name = spacetime_num_bytes_sent_to_clients_total]
-        #[help = "The cumulative number of bytes sent to clients"]
-        #[labels(txn_type: WorkloadType, db: Identity)]
-        pub bytes_sent_to_clients: IntCounterVec,
-
         #[name = spacetime_subscription_send_queue_length]
         #[help = "The number of `ComputedQueries` waiting in the queue to be aggregated and broadcast by the `send_worker`"]
         #[labels(database_identity: Identity)]
@@ -372,34 +367,6 @@ metrics_group!(
         #[labels(db: Identity)]
         pub total_outgoing_queue_length: IntGaugeVec,
 
-        #[name = spacetime_replay_total_time_seconds]
-        #[help = "Total time spent replaying a database upon restart, including snapshot read, snapshot restore and commitlog replay"]
-        #[labels(db: Identity)]
-        // We expect a small number of observations per label
-        // (exactly one, for non-replicated databases, and one per leader change for replicated databases)
-        // so we'll just store a `Gauge` with the most recent observation for each database.
-        pub replay_total_time_seconds: GaugeVec,
-
-        #[name = spacetime_replay_snapshot_read_time_seconds]
-        #[help = "Time spent reading a snapshot from disk before restoring the snapshot upon restart"]
-        #[labels(db: Identity)]
-        pub replay_snapshot_read_time_seconds: GaugeVec,
-
-        #[name = spacetime_replay_snapshot_restore_time_seconds]
-        #[help = "Time spent restoring a database from a snapshot after reading the snapshot and before commitlog replay upon restart"]
-        #[labels(db: Identity)]
-        pub replay_snapshot_restore_time_seconds: GaugeVec,
-
-        #[name = spacetime_replay_commitlog_time_seconds]
-        #[help = "Time spent replaying the commitlog after restoring from a snapshot upon restart"]
-        #[labels(db: Identity)]
-        pub replay_commitlog_time_seconds: GaugeVec,
-
-        #[name = spacetime_replay_commitlog_num_commits]
-        #[help = "Number of commits replayed after restoring from a snapshot upon restart"]
-        #[labels(db: Identity)]
-        pub replay_commitlog_num_commits: IntGaugeVec,
-
         #[name = spacetime_module_create_instance_time_seconds]
         #[help = "Time taken to construct a WASM instance or V8 isolate to run module code"]
         #[labels(db: Identity, module_type: HostType)]
@@ -410,75 +377,6 @@ metrics_group!(
         // so I'm making up some buckets based on what I imagine are the upper and lower bounds of plausibility.
         #[buckets(0.001, 0.005, 0.01, 0.05, 0.1, 0.5, 1, 5, 10, 50, 100)]
         pub module_create_instance_time_seconds: HistogramVec,
-
-        #[name = spacetime_snapshot_creation_time_total_sec]
-        #[help = "The time (in seconds) it took to take and store a database snapshot, including scheduling overhead"]
-        #[labels(db: Identity)]
-        // Snapshot creation should take in the order of milliseconds,
-        // but log data suggests that there are outliers.
-        // So let's track a wide range of buckets to get a better picture.
-        //
-        // We also track the timing without `asyncify` scheduling overhead
-        // (`snapshot_creation_time_inner`), and the snapshot compression
-        // timing with / without scheduling overhead (`snapshot_compression_time_total`
-        // and `snapshot_compression_time_inner`, respectively).
-        //
-        // Compression may have contributed to observed outliers, but is no
-        // longer included in the snapshot creation timing.
-        #[buckets(0.0005, 0.001, 0.005, 0.01, 0.1, 1.0, 5.0, 10.0)]
-        pub snapshot_creation_time_total: HistogramVec,
-
-        #[name = spacetime_snapshot_creation_time_inner_sec]
-        #[help = "The time (in seconds) it took to take and store a database snapshot, excluding scheduling overhead"]
-        #[labels(db: Identity)]
-        #[buckets(0.0005, 0.001, 0.005, 0.01, 0.1, 1.0, 5.0, 10.0)]
-        pub snapshot_creation_time_inner: HistogramVec,
-
-        #[name = spacetime_snapshot_creation_time_fsync_sec]
-        #[help = "The time (in seconds) it took to fsync a database snapshot, excluding scheduling overhead"]
-        #[labels(db: Identity)]
-        #[buckets(0.0005, 0.001, 0.005, 0.01, 0.1, 1.0, 5.0, 10.0)]
-        pub snapshot_creation_time_fsync: HistogramVec,
-
-        #[name = spacetime_snapshot_compression_time_total_sec]
-        #[help = "The time (in seconds) it took to do a compression pass on the snapshot repository, including scheduling overhead"]
-        #[labels(db: Identity)]
-        // Not sure what range to expect, but certainly slower than snapshot
-        // creation.
-        #[buckets(0.001, 0.01, 0.1, 1.0, 5.0, 10.0)]
-        pub snapshot_compression_time_total: HistogramVec,
-
-        #[name = spacetime_snapshot_compression_time_inner_sec]
-        #[help = "The time (in seconds) it took to do a compression pass on the snapshot repository, excluding scheduling overhead"]
-        #[labels(db: Identity)]
-        #[buckets(0.001, 0.01, 0.1, 1.0, 5.0, 10.0)]
-        pub snapshot_compression_time_inner: HistogramVec,
-
-        #[name = spacetime_snapshot_compression_time_per_snapshot_sec]
-        #[help = "The time (in seconds) it took to compress a single snapshot"]
-        #[labels(db: Identity)]
-        #[buckets(0.001, 0.01, 0.1, 1.0, 5.0, 10.0)]
-        pub snapshot_compression_time_single: HistogramVec,
-
-        #[name = spacetime_snapshot_compression_skipped]
-        #[help = "The number of snapshots skipped in a single compression pass because they were already compressed"]
-        #[labels(db: Identity)]
-        pub snapshot_compression_skipped: IntGaugeVec,
-
-        #[name = spacetime_snapshot_compression_compressed]
-        #[help = "The number of snapshots compressed in a single compression pass"]
-        #[labels(db: Identity)]
-        pub snapshot_compression_compressed: IntGaugeVec,
-
-        #[name = spacetime_snapshot_compression_objects_compressed]
-        #[help = "The number of snapshot objects compressed in a single compression pass"]
-        #[labels(db: Identity)]
-        pub snapshot_compression_objects_compressed: IntGaugeVec,
-
-        #[name = spacetime_snapshot_compression_objects_hardlinked]
-        #[help = "The number of snapshot objects hardlinked in a single compression pass"]
-        #[labels(db: Identity)]
-        pub snapshot_compression_objects_hardlinked: IntGaugeVec,
 
         #[name = spacetime_subscription_rows_examined]
         #[help = "Distribution of rows examined per subscription query"]
@@ -496,12 +394,6 @@ metrics_group!(
         #[help = "Total number of subscription queries by scan strategy"]
         #[labels(db: Identity, scan_type: str, table: str, unindexed_columns: str)]
         pub subscription_queries_total: IntCounterVec,
-
-        #[name = spacetime_durability_blocking_send_duration_sec]
-        #[help = "Latency of blocking sends in request_durability (seconds); _count gives the number of times the channel was full"]
-        #[labels(database_identity: Identity)]
-        #[buckets(0.001, 0.01, 0.1, 1.0, 10.0)]
-        pub durability_blocking_send_duration: HistogramVec,
     }
 );
