+++ crates/engine/LICENSE
@@ -1,0 +1,1 @@
+../../licenses/BSL.txt
