+++ crates/engine/src/db/persistence.rs
@@ -10,7 +10,8 @@ use spacetimedb_durability::{DurabilityExited, TxOffset};
 use spacetimedb_paths::server::ServerDataDir;
 use spacetimedb_snapshot::DynSnapshotRepo;
 
-use crate::{messages::control_db::Database, util::asyncify};
+use crate::util::asyncify;
+use spacetimedb_lib::Identity;
 use spacetimedb_runtime::Handle;
 
 use super::{
@@ -84,6 +85,13 @@ pub type Durability = dyn spacetimedb_durability::Durability<TxData = Txdata>;
 /// configured or the database is in follower state.
 pub type DiskSizeFn = Arc<dyn Fn() -> io::Result<SizeOnDisk> + Send + Sync>;
 
+#[derive(Clone, Copy, Debug)]
+pub struct Database {
+    pub id: u64,
+    pub database_identity: Identity,
+    pub owner_identity: Identity,
+}
+
 /// Persistence services for a database.
 pub struct Persistence {
     /// The [Durability] to use, for persisting transactions.
@@ -109,9 +117,9 @@ impl Persistence {
         durability: impl spacetimedb_durability::Durability<TxData = Txdata> + 'static,
         disk_size: impl Fn() -> io::Result<SizeOnDisk> + Send + Sync + 'static,
         snapshots: Option<SnapshotWorker>,
-        runtime: tokio::runtime::Handle,
+        runtime: Handle,
     ) -> Self {
-        Self::new_with_runtime(durability, disk_size, snapshots, Handle::tokio(runtime))
+        Self::new_with_runtime(durability, disk_size, snapshots, runtime)
     }
 
     pub fn new_with_runtime(
@@ -175,7 +183,7 @@ impl Persistence {
 /// A persistence provider is a "factory" of sorts that can produce [Persistence]
 /// services for a given replica.
 ///
-/// The [crate::host::HostController] uses this to obtain [Persistence]s from
+/// The host controller uses this to obtain [Persistence]s from
 /// an external source, and construct [relational_db::RelationalDB]s with it.
 ///
 /// This is an `async_trait` to allow it to be used as a trait object.
@@ -215,15 +223,16 @@ impl LocalPersistenceProvider {
 #[async_trait]
 impl PersistenceProvider for LocalPersistenceProvider {
     async fn persistence(&self, database: &Database, replica_id: u64) -> anyhow::Result<Persistence> {
+        let database_identity = database.database_identity;
         let replica_dir = self.data_dir.replica(replica_id);
         let snapshot_dir = replica_dir.snapshots();
         let runtime = Handle::tokio_current();
 
-        let database_identity = database.database_identity;
-        let snapshot_worker =
-            asyncify(move || relational_db::open_snapshot_repo(snapshot_dir, database_identity, replica_id))
-                .await
-                .map(|repo| SnapshotWorker::new(repo, snapshot::Compression::Enabled, runtime.clone()))?;
+        let snapshot_worker = asyncify(&runtime, move || {
+            relational_db::open_snapshot_repo(snapshot_dir, database_identity, replica_id)
+        })
+        .await
+        .map(|repo| SnapshotWorker::new(repo, snapshot::Compression::Enabled, runtime.clone()))?;
         let (durability, disk_size) = relational_db::local_durability_with_options(
             replica_dir,
             runtime.clone(),
@@ -232,11 +241,12 @@ impl PersistenceProvider for LocalPersistenceProvider {
         )
         .await?;
 
-        tokio::spawn(relational_db::snapshot_watching_commitlog_compressor(
+        runtime.spawn(relational_db::snapshot_watching_commitlog_compressor(
             snapshot_worker.subscribe(),
             None,
             None,
             durability.clone(),
+            runtime.clone(),
         ));
 
         Ok(Persistence {
