+++ crates/engine/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod db;
+pub mod error;
+pub mod metrics;
+pub mod rls;
+mod sql;
+pub mod util;
+
+pub use spacetimedb_lib::identity;
+pub use spacetimedb_lib::Identity;
+pub use spacetimedb_sats::hash;
