+++ crates/engine/testdata/v1.2/replicas/22000001/snapshots/00000000000000000000.snapshot_dir/00000000000000000000.snapshot_bsatn
