+++ crates/engine/testdata/v1.2/replicas/22000001/snapshots/00000000000000000000.snapshot_dir/objects/19/30ce81246a4cdc25e9024ae0065d053adb2efbe1b5b7af457331d330e481e8
