+++ crates/engine/testdata/v1.2/replicas/22000001/snapshots/00000000000000000000.snapshot_dir/objects/41/bb11b6d2cdc488192ee70d8175307d6f205756ed163f4237c6cba2936798dc
