+++ crates/engine/testdata/v1.2/replicas/22000001/snapshots/00000000000000000000.snapshot_dir/objects/45/4d2e2c62ff5d46c5b3e6de72d6277eb285fc2d6b0a5ac6f92498e08a9e5ecc
