+++ crates/engine/testdata/v1.2/replicas/22000001/snapshots/00000000000000000000.snapshot_dir/objects/62/22df0e5ca93d3fb22762e12161246a1d5917c61ada5d81b8dcce12fd5780b3
