+++ crates/engine/testdata/v1.2/replicas/22000001/snapshots/00000000000000000000.snapshot_dir/objects/79/4dced5633eca2ffee784d471f5203209169321083ef99de254ad24af0f6d5a
