+++ crates/snapshot/Cargo.toml
@@ -6,10 +6,15 @@ rust-version.workspace = true
 license-file = "LICENSE"
 description = "Low-level interfaces for capturing and restoring snapshots of database states"
 
+[features]
+default = ["tokio"]
+tokio = ["spacetimedb-durability/tokio"]
+simulation = ["spacetimedb-durability/simulation"]
+
 [dependencies]
 spacetimedb-table.workspace = true
 spacetimedb-data-structures.workspace = true
-spacetimedb-durability.workspace = true
+spacetimedb-durability = { path = "../durability", default-features = false }
 spacetimedb-lib.workspace = true
 spacetimedb-sats = { workspace = true, features = ["blake3"] }
 spacetimedb-primitives.workspace = true
